Dynamic arithmetic read Int boxes as raw float bits — in two places at once
test_dyn_op under the whole-module JIT printed sub=2.47e-323 mul=0
mod=4.9e-324 where the interpreter said 5, 14, 1. The values are the
tell: 2.47e-323 is the f64 whose BITS are 5, division alone was exact
(the 2^-1074 scales cancel), and multiplication underflowed to zero.
Dynamic subtraction lowers to SafeCast(HDYN->HF64) on each operand, so
an Int box was being read as a denormal. Two independent defects both
produced that read:

- jit/function.rs SafeCast unbox: loaded vdynamic.v at offset 8 with
  the DESTINATION's LLVM type, no coercion. Right only when the box's
  runtime kind is the destination kind; an HDYN register can hold any
  numeric box. Now emits a call to the dyn-cast helpers (new std
  exports hlp_dyn_toint/tofloat/todouble/toi64, which pass the dyn
  type singleton so emitted code never materializes a type pointer),
  then narrows to the register width. Null still yields zero.

- std cast.rs: hlp_dyn_castd's HI32 arm read `*(data as *mut f64)`
  and hlp_dyn_castf's read `*(data as *mut f32)` — the bit-pun itself,
  wrong for every caller, not just the JIT. Both now convert:
  `*(data as *mut i32) as f64/f32`. casti and casti64 were correct.

add= and mixed= were never wrong because Dynamic `+` compiles to the
bytecode type-switch helper, not to SafeCast+FSub — which is what made
the failure look selective.

Also gates the unconditional "[ash] hlp_throw: kind=..." line behind
ASH_TRACE_THROW (read once, gc.rs-style): the interpreter throws
through its own machinery, so the line broke every jit-vs-interp
output diff that exercised exceptions.

Verified: test_dyn_op prints Checksum: 802 in all three modes,
matching the interpreter; corpus jit-vs-interp sweep is down from 6
diffs to 1 (mandelbrot_small, the documented FMA-contraction
divergence); parity and stdlib matrices 3/3; workspace clippy clean.

+++ std/src/cast.rs
@@ -174,7 +174,7 @@ pub unsafe extern "C" fn hlp_dyn_castf(data: *mut c_void, t: *mut hl_type) -> f3
     match (*t).kind {
         hl::hl_type_kind_HUI8 => *(data as *mut u8) as f32,
         hl::hl_type_kind_HUI16 => *(data as *mut u16) as f32,
-        hl::hl_type_kind_HI32 => *(data as *mut f32),
+        hl::hl_type_kind_HI32 => *(data as *mut i32) as f32,
         hl::hl_type_kind_HI64 => *(data as *mut i64) as f32,
         hl::hl_type_kind_HF32 => *(data as *mut f32),
         hl::hl_type_kind_HF64 => *(data as *mut f64) as f32,
@@ -233,7 +233,7 @@ pub unsafe extern "C" fn hlp_dyn_castd(data: *mut c_void, t: *mut hl_type) -> f6
     match (*t).kind {
         hl::hl_type_kind_HUI8 => *(data as *mut u8) as f64,
         hl::hl_type_kind_HUI16 => *(data as *mut u16) as f64,
-        hl::hl_type_kind_HI32 => *(data as *mut f64),
+        hl::hl_type_kind_HI32 => *(data as *mut i32) as f64,
         hl::hl_type_kind_HI64 => *(data as *mut i64) as f64,
         hl::hl_type_kind_HF32 => *(data as *mut f32) as f64,
         hl::hl_type_kind_HF64 => *(data as *mut f64),
